Added BreakTimes, Tolerance. Changed DefuseTime, BombBlock.
Some defaults like the BombBlock was changed to TNT.
And DefuseTime was changed to 35.

This created a problem:
* i was able to break TNT x35 in 8.6 seconds.
* However, it took another test over 15 seconds.

So this is a definite flaw in the design
that the BreakTime for X blocks isn't the same for all players.

Probably due to network or latency issues.

The next version of config.yml will NOT have
* Tolerance:
* BreakTimes:

In this version, DefuseTime is the number of times
that you have to break the BombBlock to defuse it.

Next version, it'll be the number of seconds required
to defuse the bomb.