[Untested] hides FakePlayer in database +
added new methods:
getPlantSuccess() returns WLT.WIN
getPlantFailure() returns WLT.LOSS
getDefuseSuccess() returns WLT.TIE

added these methods so that I can remove the WLT import from
the other files (because this is an optional dependency and
BattleTracker.jar and it's WLT class may not exist).

Also, this adds readability and defines
WIN as a Successful Bomb Plant that has destroyed an enemy base.
LOSS as Bomb Plant that did NOT destroy a base due to being defused.
TIE as a Bomb that was Defused (given to the defuser).

Also, I tried to hide the FakePlayer 'Bombs Planted Defused'
in the database. But this is not tested.