Keeps track of BlockBreaks for each player
Unfortunately, when I tested this class
it took me 8.6 seconds to break TNT x35
while it took another tester over 15 seconds to destroy the same 35 blocks
of TNT.

This must be a latency problem.

And because the time to break X number of blocks will not be the same for
all players, I am going to have to use an alternative game-player mechanic
for defusing the bomb.

In future version, the defusal mechanic will be the same for planting the
bomb: InventoryOpenEvent

So the first test for InventoryOpenEvent will be if a DetonationTimer is
running. If so, start a DefuseTimer.

if not, check and see if it's the bomb carrier at the enemy base.