Fixed getBases() and updateArenasYml()
Note that the getBases() method uses a BattleArena method that requires
BattleArena version 3.9.7.3 or newer.

The getBases() method reads from bases.yml

The updateArenasYml() method handles any changes in config.yml for the
BombBlock and automatically changes any OLD BombBlock spawns in arenas.yml
to the new BombBlock.

Added getters & setters on 509.

With the getters and setters, fields will be changed from public to
private in future versions.