Moved to utility package.
Not sure if i changed anything else.

But I also added comments that show when items were introduced into
Minecraft.

Which means that if you want to support Minecraft versions older than
1.5.2 then you'll need to comment out or delete those items.

Instead of hardcoding the items, I developed a loop from Material.values()
but it was only able to reconize 8 out of 13 InventoryTypes. Hopefully, I
can fix this in the future to support old AND new versions of Minecraft
(i.e. blocks that haven't been released yet and have inventories
associated with them).