Handles conversion of plugin versions to Integers for comparison.
getBAversion() - returns the 6 digit BA version with padded zeros to the
right.

getVersion(String) - Currently, functionality is limited to deleting
decimal points and creating a 6 digit Integer with zeros padded to the
right.

if necessary, future versions will support other version formats that have
different separators (like dashes, etc).

Currently, the Integer form of the version is used for comparison to what's
on the server and provide backwards compatibility.