Fixes major NPE bug.
Ocassionally, a major bug would appear when someone planted the bomb:
which started the DetonationTimer.

The problem occured when one of the teams died during the DetonationTimer.

Added a temporary try-catch block until i was able to fix the error.

This error is now fixed via 2 files:
BombArenaConfig.yml: victoryCondition: NoTeamsLeft
BombArena.java: onTeamDeathEvent()

Now, the match does NOT immediately end if an entire team dies while the
DetonationTimer is running.

If a team dies while a DetonationTimer is running, then the match will
continue until the bomb explodes or someone defuses the bomb.