Version 1.1.4 changes.
The problem occured when a team died while the DetonationTimer was
running.

The fix was with 2 files:
BombArenaConfig.yml: victoryCondition: NoTeamsLeft
BombArena.java:onTeamDeathEvent()

This event checks wether a DetonationTimer is running when a team dies.
if so, let the Timer decide the winner.
if not, declare the winners & losers.

ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss