Refactored from Comparator<> to Comparable<>
refactored since each object has the version attribute.

This version attribute is what is currently running on the server.

Example:
if (!Version.getPlugin("Abc").isCompatible("1.2.3-b777")) {
   disableMe();
}

No need to check if plugin Abc is installed or enabled()

isCompatible() and isSupported() check if the plugin is null or if the
plugin is not enabled.

In other words: If you attempt to instaniate a Version object for a plugin
that isn't even installed on the server, then then any compatibility check
is false.

DISABLED PLUGINS: You can get their version and print them out. But all
compatibility checks will fail. (this is intended because you don't need
to run version dependent code against something that is disabled).

Thanks to Tux2 for pointing out that version checks will fail for
GroupManager. Might add possible future support for GM. Should work for
the other 99% of plugins.

Server tip: check against
Version.getServer().isSupported("1.X.Y-R9.9-SNAPSHOT");