Refactored conditional
if-statement no longer checks if the plugin is null because the Version
object takes care of this internally.

Added commented-out code that will eventually support EnjinMinecraftPlugin
custom stats. The code is for testing purposes only, and will be released
in a future versions of BombArena.

The code is Enjin version dependent upon v2.6.0