v1.1.5 Fixed bug and added new plant mechanic.
NEW FEATURE: onBombPlace(): in previous versions, this would just alert
the player of "Improper Bomb Activation."

onBombPlace() now attempts to trigger onBombPlant() if the planter is
close enough to the base. The distance was hard coded during testing. It
will be configurable in future updates.

Be aware that this could potentially give an unfair advantage to the
attacking team because while the defenders will be forced to get really
close to defuse the bomb, the attackers will have the luxury of planting
from a further distance. Use this option with caution.

The intention for this option is to make it easier and intuitive for new
players to plant the bomb. New players may not initially know how to plant
the bomb, so now they'll have a 2nd way. It makes it easier to learn the
game, and takes away some of the frustration from players who cannot
figure out the classic bomb plant mechanics of interacting with the
baseblock.

BUG FIXES:
1. onBombCarrierLeave() has finally been implemented. This fixes a
situation where a player could pickup the bomb then disconnect or leave
the game... leading to a scenario where the BombArena match would NOT have
a bomb available to players who wanted to complete the objective.

2. the bomb planter no longer drops a bomb if the DetonationTimer is
running. (i.e. He has planted the bomb and therefore does not have it).

3. the plugin.carriers for the match will no longer get cleared
onBombDrop() if the planter dies after he has planted the bomb. See #2.

4. Fixed an unusual and rare error that would occur when the bomb got
planted and all the players left the match... Meaning that the error would
occur because the DetonationTimer was running while the match was
finished. onFinish() now cancels the match timers.

Hopefully, this is the last of the bug fixes so that I can focus on adding
new features like the SND rule-set. And also focus on releasing other Arenas
like HostageArena and ConquestArena.

To-Do:
Add a new config option for the BaseRadius that defines how much leeway to
give new players who plant the bomb like it's a block... And defines how
far away they can be from the plant site.