Added HolographicDisplays classes
http://dev.bukkit.org/bukkit-plugins/holographic-displays/

Since HolographicDisplays is an optional dependency, API methods that
create Holograms will return a hologramID (since we cannot return a
hologram because the Hologram class may not even exist at runtime on the
server).

Save the hologramID and use it to perform changes later (like stop
displaying it, or change the location, etc).