Deprecated
There is a good chance that this class will no longer be used.

The original use-case-scenario was a mis-configured arena that doesn't
allow item drops.

if the server Admins accidentally mis-configure the Arena to NOT allow
item drops... Then it'll break BombArena, because the bomb would only be
able to be picked up ONE TIME.

This class was used to make a PlayerDropItemEvent and call it.