Changed project's file structure to default Maven
in order to make testing easier (so I can use the src/test directory).

Most of the actual changes in this commit are in BombArena.java

Support for HolographicDisplays has been added.

onPlayerDeathEvent() no longer creates a PlayerDropItemEvent. This means
that checks need to be added to make sure that Arenas allow item drops.