Deleted HoloManager.java
The purpose of the HoloManager was to keep track of Holograms and pass out
IDs, and use those IDs later to delete Holograms.

Each implementation of the HologramInterface now does the work of passing
out IDs because (for example), the parameter for addHologram() can be
com.gmail.filoghost.holograms.api.Hologram
or
com.dsh105.holoapi.api.Hologram

So at first, it might seem like duplicate code... But really, it's more
like overloaded methods.