Added sorting, zero padding, and comments.
Bug fix:
Padding with zeros fixes a rare bug where the server is running (for
example) a plugin with version "1.2.0", and you ask isCompatible("1.2").
In this situation, the compatibility check would fail eventho is
shouldn't.

I thought of this because BombArena version 1.2.0 is about to be released.

Got Permission from DSH105 to use some of his ideas from:
com.dsh105.commodus.Version;

The article that I read on Comparator/Comparable said to return one,
negative one, and zero... But I really like how DSH105 simply returns the
difference: very elegent.

compareTo():
Also, someone in IRC told me that I should compareTo() other Version
objects instead of String objects. Initially, I didn't understand the
use-case-scenario, but then I figured out that it could be used for
sorting. Eventho, I'll probably never use this particular functionality, I
changed it... just in case anyone else wants to use it, (or even me in the
future).

Technically, I don't think the Version object should know about Plugins or
Bukkit Servers, but it is a little convenient for isCompatible() and
isSupported() to check for me if a plugin is even enabled.

So I might change this in the future... along with the factory methods...
Those should probably be in their own class, where the Version object
knows nothing about Plugins or Servers, but the VersionFactory does know
about them and how to use them to conveniently construct a Version object
for the user.