Added tests for sorting and compareTo().
compareTo() no longer returns one, negative one, and zero.
It now returns the difference: which can be positive, negative, or zero.

So I added some tests to make sure this is working as intended.

compareTo() now accepts a Version object instead of a String.

This allows for sorting, so I added a test for this.

testSort() :
This method will generate 25 random version and print them out.

It will then sort them in ascending order.
It will verify that the next one in the list is greater than or equal to the
current index.

It will then sort them in descending order.
It will verify that the next one in the list is less than or equal to the
current index.