Fixed holograms for dropped items.
Previously, there was a bug where the hologram for a dropped item could
potentially glitch into the ground, and display improperly.

The fix was on line 316 to add a Runnable() that would follow the item as
it's thrown thru the air... And the Runnable() will stop when the bomb
reaches the ground.

Also refactored the hd() method call to holograms() method.