Finished Search And Destroy.
SndArena is tested and fully working.

Attacking Team:
The team that spawns closest to the bomb is assigned as the attacking
team.
The remaining team is assign as the defending team.

To-do:
Right now, the base hologram is put on the closest base to the bomb
carrier (instead of both bases).