Added support for older versions of BattleArena.
BombArena now requires BattleArena v3.9.6+

The previous requirement was 3.9.8+ because of ArenaFactory.

BombArena still uses the ArenaFactory... except now, the import is in a
wrapper class to prevent ClassNotFoundException at runtime.

If the server has 3.9.8 or newer, it will registerCompetition() with an
ArenaFactory.

However, if the server is running an older version, it will
registerCompetition() without an ArenaFactory.

It will register with BombArena.class & SndArena.class