Added MatchUtil::setTime()
The problem is that BattleArena does NOT have a way to change the match
time. Which I can see why: because needed to change the match time is
extremely rare. Why would someone need to do this ?

With the Sabotage & Search-N-Destroy rule-sets, this exact scenario comes
into play... When the bomb is planted, the match time needs to change to
whatever the DetonationTimer for the bomb is.

There are many possible solutions to this problem:
1. Add an API method in BattleArena that allows changing of time.
2. Hack the internals and set time.

The downside of an API method would be that BombArena would then require
the BattleArena version with that API method (or higher). So BombArena
would no longer be backwards compatible with BattleArena.

Unfortunately, the downside of using this MatchUtil::setTime() is that it
is tightly-coupled to the inner workings of BattleArena (which is also
bad), because eventho with this approach, we have backwards compatibility,
it is not guarenteed to work in future versions.

Therefore, a hybrid approach is probably best: Have an abstraction of
MatchUtil.setTime() with different implementations depending on what
version of BattleArena the server is running. If it's running the new
version, then merely forward any setTime() calls to the appropriate
BattleArena API method. If it's running an older version, then use the
hacky, tightly-coupled approach.