Match Time now correctly changed on bomb plant.
Anytime the bomb is planted, the match time will now be changed to
whatever the DetonationTimer is.

For example: The DetontationTimer is 45 seconds. The PlantTimer is 6
seconds. If a player starts planting at 16 seconds, then on a success bomb
plant at 10 seconds, the Match Time will change from 10 to 45... including
what is displayed on the Scoreboard.