Fixes read/write/update of old config.yml
Previously, the problem was that an old config.yml wouldn't have the new
nodes. This was caused by using Bukkit's API getConfig().contains()
instead of getConfig().isSet().

For config handling, the way it works now is

saveDefaultConfig();
// This only writes config.yml if it doesn't exist.

addAllNodes();
// add new nodes to an old config.yml
Minor problem of the nodes being out of order tho.

loadDefaultConfig();
// load all the nodes into memory.

restructureNodes();
// delete all the nodes and add them back in order.